requirements.txt
torch>=2.6.0
transformers>=4.53.1
sentence-transformers>=4.1.0
faiss-cpu>=1.11.0
PyPDF2>=3.0.1
numpy>=2.0.2 